Move top level XML parsing/formatting functions to ParserHelper
