+++ NAS2D/Configuration.cpp
@@ -22,56 +22,6 @@
 using namespace NAS2D;
 
 
-namespace {
-	template <auto fileFormatParser(const Xml::XmlElement& element) = subTagsToDictionaryMap>
-	auto parseXmlFileData(const std::string& xmlString, const std::string& sectionName = "", const std::string& requiredVersion = "")
-	{
-		Xml::XmlDocument xmlDocument;
-		xmlDocument.parse(xmlString.c_str());
-
-		if (xmlDocument.error())
-		{
-			throw std::runtime_error("Error parsing XML file on (Row " + std::to_string(xmlDocument.errorRow()) + ", Column " + std::to_string(xmlDocument.errorCol()) + "): " + xmlDocument.errorDesc());
-		}
-
-		auto* root = !sectionName.empty() ? xmlDocument.firstChildElement(sectionName) : xmlDocument.rootElement();
-		if (!root)
-		{
-			throw std::runtime_error("XML file does not contain tag: " + (!sectionName.empty() ? sectionName : "(root element)"));
-		}
-
-		if (!requiredVersion.empty())
-		{
-			const auto actualVersion = root->attribute("version");
-			if (actualVersion != requiredVersion)
-			{
-				throw std::runtime_error("Version mismatch. Expected: " + std::string{requiredVersion} + " Actual: " + actualVersion);
-			}
-		}
-
-		return fileFormatParser(*root);
-	}
-
-	template <typename Data, Xml::XmlElement* dataFormatter(const std::string& tagName, const Data& data) = dictionaryMapToElement>
-	std::string formatXmlData(const Data& data, const std::string& tagName, const std::string& comment = "")
-	{
-		Xml::XmlDocument doc;
-
-		if (!comment.empty())
-		{
-			doc.linkEndChild(new Xml::XmlComment(comment));
-		}
-		doc.linkEndChild(dataFormatter(tagName, data));
-
-		// Write out the XML file.
-		Xml::XmlMemoryBuffer buff;
-		doc.accept(&buff);
-
-		return buff.buffer();
-	}
-}
-
-
 Configuration::Configuration(std::map<std::string, Dictionary> defaults) :
 	mDefaults{std::move(defaults)},
 	mSettings{mDefaults}
